[BACKEND] Change DenseMap to std::map in Allocation (#1092)
Since we need to get the address/reference of each entry, it is not
appropriate to use DenseMap here because rehashing on DenseMap will
reallocate buckets and change addresses of entries. Unlike DenseMap,
associative containers in STL do not change references to existing
entries when other entries are inserted or deleted.


https://stackoverflow.com/questions/61771352/will-it-change-the-address-of-a-existed-keys-value-when-inserting-new-keys