Switch pre-commit clang-format to v16.0.6. (#2342)
Google uses clang-format at LLVM HEAD.  clang-format's formatting is not
stable, so we want to minimize the difference between the pre-commit
clang-format and HEAD to minimize differences with Google's formatter.

In practice, it appears that there are no relevant changes to the
formatting, so this is a nop.  :shrug:

Tested by running `pre-commit run --all-files`.