[BACKEND] Exclude arith.select from hoisting for dot_operand layout (#3031)
Hoisting dot_operand layout through arith.select crashes the compiler
due to propagating the layout to the mask input, which is of i1 type.
Also hoisting further up the mask's parents will probably be of integer
type. The crash happens during ConvertTritonGPUToLLVM pass.

2 examples where the crash happen:
- arith.cmpi -> broadcast -> layout_conv -> arith.select (dot_op layout)
-> dot.
In this example the crash would happen due to
SharedToDotOperandMMAv2.cpp assuming that the bit-width is at least 8
when computing elemBytes. It also assumes the types are mma-compatible
(for Ampere for example mma16816) which don't include int type when
doing floating dot.

- arith.cmpi -> layout_conv -> broadcast -> arith.select (dot_op layout)
-> dot.
This example requires allowing HoistLayoutConversion to also propagate
through Broadcast. Again, arith.cmpi is on integer 32 which is not
supported in the lowering of shared -> dot_operand layout.