[OPTIMIZER] Allow rematerializing through scf.if ops (#3217)
This allows the remove layout conversion pass to rematerialize values
through an `scf.if` block. This is done by adding new return types to
the if block and letting the `IfOp` canonicalization patterns remove any
resulting dead code.

Testing this also uncovered a quirk in the `ForOp` rematerialization
process. It calls `replaceForOpWithNewSignature` which eagerly replaces
all users of the pre-exiting for-loop. However, we can get the case
like:

```mlir
%1 = scf.for ... {
    %2 = scf.for ... {
        scf.yield ...
    }
    scf.yield %2
}
```

The inner for loop `%2` is rewritten before the outer yield, so when
`rewriteSlice` checks `slice.count(operand)` it can never match because
the operand has been updated to the new loop, but the slice still has
the old loop result.

I fix this by delaying the `replaceAllUsersWith` call until the very end
of `rewriteSlice`. This also allows mixtures of `scf.for` and `scf.if`
to be nested.