[FRONTEND] Keep pointer_type name and its __str__ in-sync. (#3453)
This change ensures that `name` cannot get out of sync with `__str__`
and `__repr__`, since all of the fields are technically mutable and can
be changed at any time. This also comes with a potential performance
benefit of computing this f-string only once instead of computing it on
each `__str__`/`__repr__` invocation.

It would be even better to also to make this class immutable, but it's a
separate story.