Fix the wrong clock rate unit in TFLOPS estimation for matmuls (#3556)
While investigating weird autotuned performance of some Triton kernels,
I noticed that theoretical estimates returned from
`estimate_matmul_time()` were suspiciously high. It turns out that
`nvidia-smi` provides the max SM clock rate in MHz (e.g., something like
`1410` for A100 GPUs), but `get_max_*_tflops()` functions expect
`clock_rate` to be provided in KHz (e.g., `1410000`). I don't see this
expectation to be documented anywhere, but:
* `clock_rate` is multiplied by `1e-9`
[here](https://github.com/openai/triton/blob/d988b759ef1f83632ed2f8aa99e43d6349077acf/python/triton/testing.py#L403),
which only gives you TFLOPS if `clock_rate` is in KHz
* the code in tests already accounts for this by
[multiplying](https://github.com/openai/triton/blob/d988b759ef1f83632ed2f8aa99e43d6349077acf/python/test/regression/test_performance.py#L65)
the `nvidia-smi` result by `1e3`

The end result of this is that `compute_ms`
[here](https://github.com/openai/triton/blob/d988b759ef1f83632ed2f8aa99e43d6349077acf/python/triton/ops/matmul_perf_model.py#L60),
in fact, becomes `compute_us` and therefore completely dominates the
final expression
[here](https://github.com/openai/triton/blob/d988b759ef1f83632ed2f8aa99e43d6349077acf/python/triton/ops/matmul_perf_model.py#L92).
In other words, the loading/storing times basically become irrelevant.

**Disclaimer** I actually couldn't find find a test case where the fix
would make a difference: large compute-bound matmuls are obviously not
affected, but at least on A100, there is no meaningful performance
difference even for smaller sizes I tested (which should be
memory-bound). The set of pruned configs indeed changes after the fix,
but it doesn't seem to matter because all configs are essentially the
same.

It is also possible that fix will actually degrade the performance of
the kernels which were written with the old performance model in mind.
However, I decided I'd post this PR just in case because I think that
this is definitely not the intended behavior.

Co-authored-by: Justin Lebar <justin.lebar@gmail.com>