[backend][amd] Change to dynamically load libamdhip64.so (#3646)
This commit changes the AMD backend to perform dynamic loading of the
HIP runtime library. This is better because of various reasons,
including but not limited to:

* We can be consistent with PyTorch regarding the library used
* We can give explicit control to developers as which dylib to use
* We can trim down the dependencies in Triton
* etc.

The dynamic loading procedure will try to load the `libamdhip64.so`
directory from the current process if exists. Otherwise, it goes through
the following steps:

* If we are told explicitly what HIP runtime dynamic library to use with
the `TRITON_LIBHIP_PATH` environment variable, obey that. Do not attempt
any other library locations.
* Otherwise, first search the HIP runtime dynamic library packaged with
PyTorch. It's very likely that we run Triton together with PyTorch. This
makes sure we use the same dynamic library to avoid version mismatch.
* Then try to see if developer provides a HIP runtime dynamic library
using `LD_LIBARAY_PATH`.
* Afterwards try to search the loader dynamic library resolution paths.
* As a last resort, guess if we have it in some common installation
path.

Proper verbose error messages will be given if none of the above hits.

Fixes https://github.com/openai/triton/issues/3620