Cache Triton compilation artifacts during CI. (#3680)
The vast majority of CI time is spent compiling Triton kernels. This
caches
those, so if the compiler doesn't change, we should be able to reuse the
cached
kernels. OTOH if the compiler does change, Triton will not (or, should
not)
use stale compilation artifacts, because the Triton source code is
effectively
part of the cache key.

Because the Triton C++ compiled code is part of the cache key, this will
only
work if the Triton C++ build is deterministic. We should be able to see
if
caching is working by comparing the names of the compilation artifacts
between
two nop commits to `main`.

While we're here, we also run pre-commit on changes in the main branch.
Previously we only ran it on pull requests. The reason for this is that
if we
don't run it on main, the pip caching won't work on non-main branches
(because
the cache action only pulls from the current branch or main).

Also while we're here, rewrite the AMD workflow definition to reference
the
nvidia workflow definition where they match -- it was getting hard to
keep the
two workflows in sync. Because github actions don't support YAML
references, I
had to add a preprocessing step to pre-commit.  But it's not too bad.