[backend][amd] Make target specification and usage consistent (#3717)
This commit improves how target architecture are handled in the AMD
backend.

Previously we have three ways of specifying the target: 1) the `gfx*`
target architecture, 2) a compute capability number that matches the
CDNA generation, e.g., 300/200/100, 3) a compute capability number that
tries to pretend to be CUDA equivalence, e.g., 80/90. These three
schemes are quite inconsistently used and not all the places are
properly seeing the target it should be--it can be quite confusing.

In general, for CodeGen towards AMD GPU, the `gfx*` target architecture
should be the original source of capability and feature information.
Further, for AMD, there is no exact correspondance of NVIDIA CUDA
compute capability numbers.

So this commit changes to base on the `gfx*` target, and adds utility
functions to deduce features and ISA families from it. All existing
passes and patterns in the AMD backend are updated to properly use such
information. "Compute capability numbers" are removed entirely.

Along the way, also updated TritonToTritonGPUPass to take the compute
capability number as optional. We should make it more backend agnostic
but that can happen later.