[BACKEND][Pipeliner]] Handle masked load with non-zero other values (#3758)
Currently pipelining a masked load doesn’t deal with non-zero `other`
values, as it falls back to `cp.async` which fills the destination with
zeros when the mask evaluates to `false`. I'm fixing this by issuing an
extra select instruction between the loaded values from SMEM and the
`other` values.