Fixing LUT pipeline regression (#3784)
After the refactoring of shmem Ops we have introduced the regression in
the indirect loads pipelining. Waits were scheduled in the wrong stages,
totally independent of the local_loads, resulting in incorrect schedule
being generated. This PR introduces a token returned from the async_wait
op, consumed by the `local_load`, so the pipeliner can correctly
discover the dependency between those two ops, and schedule waits in the
correct stage with the `local_load`.