[PIPELINING] Introducing coarse schedule, schedule `if`s to maximize instruction scheduling opportunities in the loop (#3744)
This change introduces the concept of coarse (high level) schedule. It
is being created by assigning interesting ops in the loop the stage and
the ordering cluster. Ops with higher ordering cluster number are
scheduled in the loop after the ops with lower ordering cluster. After
assigning key instructions to stage and cluster, rest of the
instructions are assigned automatically to create a valid schedule.
This new mechanism is used to push out the prolog and epilog `if`s to
the boundaries of the loop, to keep the central block of the loop as big
as possible to expose instruction scheduling opportunities.