[PROTON] Support setting scope/op properties (#3865)
Different from "metrics" that can be aggregated, "properties" in proton
cannot.

This is particularly useful for fine-grained IR analysis for GPU
kernels. We need to know the file path of each kernel to get the path to
IR files. Using the existing approach, a file path has to be appended to
a scope name, which is not something that the user wants to know.
Instead, by setting a `file_path` property, the path will be hidden from
users. Given that each triton kernel corresponds to a unique file path,
the solution should work.