[INTERPRETER] Fix scalar mask when the value is False  (#3880)
We hacked scalar tensor's `__bool__` method so that `if (scalar tensor)`
can be executed in the interpreter mode.


https://github.com/openai/triton/blob/main/python/triton/runtime/interpreter.py#L700

But scalar tensors shouldn't be evaluated in `semantic.py`.

We could have `tl.load(..., mask=tl.tensor([False]),
other=tl.tensor([1.0])`. It is still valid. Only when `mask=None,
other!=None` it is invalid.

Without this PR, `not tl.tensor([False])` is evaluated as True and
raises the wrong `ValueError`.

So we should instead check if the tensor is None or not.