[AMD] Fix an issue with deduplication with mfma layout (#3870)
This PR fixes the following two issues
1. `getContigPerThread()` for sliceLayout should erase the `dim` after
getting the vector from its parent
2. `maybeDeduplicate()` should use `getContigPerThread()` instead of
`getSizePerThread()`

Note that `sizePerThread` and `contigPerThread` are the same for blocked
layout and sliceLayout with blocked layout as parent. But they are
different for mfma layout. Using sizePerThread in maybeDeduplicate
causes an issue when we do masked store with mfma layout. Because the
`constancy` is not adjusted properly, mask values that are not supposed
to be duplicated are duplicated, which leads to out-of-bound memory
access.

---------

Co-authored-by: Lixun Zhang <lixun.zhang@amd.com>