Tiny change: Added/Improved error message in visit_For (#4171)
While developing a kernel, I was given the error message
"AssertionError()" without much helpful context on how to proceed with
debugging. I could only solve it by understanding that part of the
triton source code and spending half a day.

That's why I'm (1) adding an error message to this part of the code, and
(2) making the error message above it clearer (like it is in
visit_While).

This should allow the end user to debug this error without the need to
dive into the triton source code.