setup.py: introduce TRITON_OFFLINE_BUILD (#4414)
In e.g. Nixpkgs (probably in other similar distributions as well) it's
important to be able to inject dependencies rather than use
vendored/locked dependencies. When we fail to provide complete
dependencies for triton we want to get an immediate error, rather than
an attempt to download a substitute from the internet (which fails after
a timeout anyway, because we the actual builds offline). Sometimes we'd
also try to override a recipe with a different version of a dependency
than declared upstream, so e.g. we tend to treat strict semantic version
checking as an optional feature.

This PR proposes an additional boolean environment variable meant to
disable vendoring and fetching attempts in favour of either an immediate
error ("input declared but not defined") or a later compile/link time
error.

I'm happy to maintain a patch on the Nixpkgs side as well if you think
this doesn't belong in your repo.

I'm waiting for the builds to finish.

The core Triton is a small number of people, and we receive many PRs
(thank
you!).  To help us review your code more quickly, **if you are a new
contributor (less than 3 PRs merged) we ask that you complete the
following
tasks and include the filled-out checklist in your PR description.**

Complete the following tasks before sending your PR, and replace `[ ]`
with
`[x]` to indicate you have done them.

- [x] I am not making a trivial change, such as fixing a typo in a
comment.

- [x] I have written a PR description following these
  [rules](https://cbea.ms/git-commit/#why-not-how).

- [ ] I have run `pre-commit run --from-ref origin/main --to-ref HEAD`.

- Select one of the following.
  - [ ] I have added tests.
    - `/test` for `lit` tests
    - `/unittest` for C++ tests
    - `/python/test` for end-to-end tests
  - [ ] This PR does not need a test because `FILL THIS IN`.

- Select one of the following.
  - [ ] I have not added any `lit` tests.
- [ ] The `lit` tests I have added follow these [best
practices](https://mlir.llvm.org/getting_started/TestingGuide/#filecheck-best-practices),
including the "tests should be minimal" section. (Usually running Python
code
    and using the instructions it generates is not minimal.)