[Frontend] Fix codegen when top level control flow occurs after an unconditional return (#4885)
Fixes #4883

In the issue we have a constexpr if condition which gets inlined in the
frontend. Since the if contains a return statement, this means we
generate a top-level return with dead code after it. As part of the code
generation, we clean this up by removing dead code after a return
statement in any given basic block otherwise LLVM complains that there
are statements after the block terminator.

However, this code also has complex control flow with multiple basic
blocks in the dead code region. The result is we delete dead operations
which still have users in the other unreachable basic blocks.

Here I fix the issue by creating a new basic block after any return
statement, and setting the insertion point to that. This means we don't
need the dead code cleanup step and can instead rely on the optimizer to
remove any unreachable blocks.